Issue: Use date picker #7
Modified `app/component/search.js`  to use the HTML5 date picker. Added code to take the resulting date string and remove the hyphens before using it in the NYT query. Also using today's date to be used as default for the end date input, and used today's date to set the maximum for the start and end date inputs.